fix create storageclass notify message can't not get item.getId() problem
