--- client/components/+storage-classes/config-storageclass-dialog.scss
@@ -1,3 +1,0 @@
-.ConfigStorageClassDialog {
-  width: 700px;
-}
